project02/requirements.txt
streamlit
numpy
flask
nltk==3.9.1 
pandas==2.2.3 
matplotlib==3.10.0 
textblob==0.18.0.post0 
vaderSentiment==3.3.2 
transformers==4.47.1 
scikit-learn==1.6.0 
gensim==4.3.3 
seaborn==0.13.2 
torch==2.5.1 
ipywidgets==8.1.5
scipy
spacy==3.7.5
thinc==8.2.5
together
python-dotenv
langchain-together
langchain_core
langchain_community
langchain_openai
langchain_unstructured
pypdf
pinecone-notebooks 
pinecone-client==3.2.2
chromadb
fastapi